requirements.txt
chatterbox-tts>=0.1.2
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
gradio>=4.0.0
torch>=2.0.0
torchaudio>=2.0.0
librosa>=0.10.0
soundfile>=0.12.0
psutil>=5.9.0
nvidia-ml-py>=12.0.0